Added Exception details
The "app.UseDeveloperExceptionPage" doesn't care so much about the
incoming requests.
It's call the next piece of middleware. It's wait to see if anything
later in the pipeline generates an exception.
And if there is an exception, this piece of middleware will give us an
error page with some info. This information can be very useful for the
developers.