Update controllers and actions.
It's common to derive a controller from a controller base class provided
in "Microsoft.AspNet.Mvc" namespace.
This base class give us access to lots of info about a request. As well
a methods help me to build results to send back to client. So I inherit
this base class in my controllers. After that we have a bigger API.
They're a lot of properties that represent contextual info.

The advantage of using IActionResult is that it's a formal way to
encapsulate the decesion of controller. So the controller decides what
to return (string, HTML, JSON). This give us more flexability and makes
controllers a little bit more testable.

Added the restaurant model.
Show restaurant info in the home page.