Added a service that simulates data access
We will use the database later and we want to display a collection of
restaurants. So we add a service/component that simulates data access.
Added a component that a controller can use to access restaurant
information and the controller doesn't directly instantiate restaurants
itself. We use "IRestaurantData" interface in the controllers and the
controllers don't know that I'm actually using "InMemoryRestaurantData".
This raises the abstraction.
Set up the controller to program against IRestaurantData.
Added some fakes restaurants and show them in the home page.

NOTES:
Anytime you use something that is just a list of something, list isn't
thread safe, so you have to be a little bit careful using it in a web
application. But I'm going to assume we only use InMemoryRestaurantData
when developing or when testing, so we're not going to have to deal with
multiple users. When we actually deploy the application, we'll use a
real SQL Server database. And a database can handle multiple users
easily.