Accept form input to create new restaurant
Added RestaurantEditViewModel. We use this model instead the
"Restaurant" entity to receive inputs from the form. We use that to
don't open any secure holes. The view model will contain only the
properties we expect to receive in the HTTP request.

Changed constructor and private list of restaurant in
InMemoryRestaurantData to static because we want to always work with the
same list throughout the life of the application. It doesn't handle
concurrent users!

Notes: If my restaurant has any properties or any attributes that I
don't want network data to get into, then I am much better off not using
an entity but using an input view model. It's much safer.