Use the POST-REDIRECT-GET pattern
Use the POST-REDIRECT-GET pattern. Now after the user creates a new
restaurant, he can bookmark the page with details about this restaurant.

Before this commit the URL after new restaurant created was /Home/Create
now is /Home/Details/{restaurant.id}.
We fixed this. This removes the danger to duplicate restaurants when to
refresh the browser after we created the new restaurant.