Implemented a DbContext
Added an OdeToFoodDbContext that will give us access to a database.
Added a SqlRestaurantData class which we will use in our controllers to
talk to the database.

Notes: To use the Entity Framework, we need to create a class that
derives from the Entity Framework DBContext base class. Each DBContext
class that you write will give you access to a single database.