Added Entity Framework Migrations
Added a migration.
Apply the migration to update a database.

To achieve this thing, I used the Developer Command Prompt for VS2015.

First, we use "dnvm" (.Net Version Manager), after that we use "dnvm
list". This command shows you your runtimes.
Next, we need to select one from this runtimes, we use the command "dnvm
use {name of the Runtime}" and that will give us access to the dnx (.NET
Execution Environment). If we include the "-p" after the name of the
runtime, then we won't have to execute this command again unless we want
to change the runtime that we using.
Using dnx, we can call into commands that we have listed in the
project.json file.

After that, we can just type "dnx ef" that will show us info about the
EntityFramework.Commands.
We can type "dnx ef migrations" and we will get a help screen again.
Next, we execute the "dnx ef migrations add {the name of migration}".
This command will generates the first migration or add new migration if
we already have old migrations. And by default, the code is generated in
"Migrations" folder.

Next, we can update the database with using the "dnx ef database
update".
At the first time, when we don't have a database yet, this command will
generate the database.
This command applying the last migration.

Notes:
One way to get a database set up is to use Entity Framework to create a
database.
EntityFramework.Commands will change at some point in the future.