Make the layout default for every view.
Added the _ViewStart.cshtml file to use the layout for all the views
instead to tell in every view which is the layout.
Remove the duplicated code in the Create and Details views. Now we use
the layout for these views too.

Notes: The Razor view engine in MVC has a convention where it will look
for any file with the name _ViewStart.cshtml and execute the code inside
of this file before executing the code inside of an individual view. The
code inside of a ViewStart file cannot render into the HTML output of a
page, but it can be used to remove the duplicate code from the code
blocks inside of the individual views.