Added the _ViewImports.cshtml file
Notes: In addition to the ViewStart file, there also a ViewImports file
that the MVC framework will look for when rendering any view. Like the
ViewStart file, I can drop ViewImports.cshtml into a folder, and the
ViewImports file can influence all the views in the folder hierarchy.
This view is new for this version of MVC, and here's the idea. In
previous versions of MVC, we could use an XML configuration file to
configure certain aspects of the Razor view engine. Those XML files are
gone. I use code instead. The ViewImports file is a place where I can
write code and place common using directives to pull in namespaces that
my views need.