Added the edit action
Added the edit action in the application.
Use tag helpers into the edit view.
Added "Commit" method in IRestaurantData and implemented in classes
which inherit that interface.

Notes: The Entity Framework doesn't flush anything to the dabatase until
we call SaveChanges. We have the ability to pull a restaurant out of the
database, modify some of the properties, and then just ultimately tell
my DBContext object to save changes. It knows what restaurants have been
pulled out of the database, and it knows which properties have been
changed. We don't need to tell it to explicitly update or save a
restaurant.