Enable the client-side validation
Added the three script files: jQuery, the jQuery validation plugin, and
the Microsoft script is known as jQuery Validation Unobtrusive.

Notes: The whole idea behind data- attributes is that they're custom for
your application, and your application will have to provide some
JavaScript to act on those attributes. Fortunately, Microsoft provides
some JavaScript that we can use to consume this validation metadata and
enforce validation on the client, the script that Microsoft includes
built on top of a popular jQuery plugin known as jQuery Validation.

jQuery has to come first before any plugins.