Used CDNs and fallbacks
After the changes the application use "node_modules" for development and
then switch over to use a CDN when we put the app into testing or
staging or production.

Notes: A CDN can typically deliver script files faster than your own
local servers.
This is a CDN run by Microsoft: http://asp.net/ajax/cdn. On this page,
we can see the list of all libraries that are hosted on Microsoft CDN.